tbh_secure_agents/security_validation/requirements.txt
# Core dependencies
numpy>=1.20.0
scikit-learn>=1.0.0

# LLM dependencies
google-generativeai>=0.3.0

# Optional dependencies
matplotlib>=3.5.0  # For visualization
pandas>=1.3.0      # For data handling
